Introduce the repo manifest xml files
The Android R&D, which we are developing is built with the help of the
repo tool. Repo is based on usage of the xml manifest files.

This patch is introducing several xml repo manifest files, which
contain links to all required projects that are needed to build the
Android 12 using the 'xenvm-trout' device.

The logic behind the introduced files is the following one:

- default.xml - entry point to include other files and locate the common
declarations
- android-12.1.0_r8.xml - the original Android repo manifest, which was
fetched from its original location
- android-12.1.0_r8-override.xml - anything we add on top of the
original manifest
- aosp-device-xenvm-trout.xml - anything related to the 'xenvm-trout'
Android device

Signed-off-by: Vladyslav Goncharuk <vladyslav_goncharuk@epam.com>